[general] Merge Forg.interval and Fog.timeout into a single file.
