+++ lib/fog/core.rb
@@ -20,7 +20,6 @@ require 'fog/core/current_machine'
 require 'fog/core/deprecation'
 require 'fog/core/errors'
 require 'fog/core/hmac'
-require 'fog/core/interval'
 require 'fog/core/logger'
 require 'fog/core/model'
 require 'fog/core/mock'
@@ -29,8 +28,8 @@ require 'fog/core/service'
 require 'fog/core/ssh'
 require 'fog/core/scp'
 require 'fog/core/time'
-require 'fog/core/timeout'
 require 'fog/core/wait_for'
+require 'fog/core/wait_for_defaults'
 require 'fog/core/class_from_string'
 require 'fog/core/uuid'
