+++ lib/fog/core/wait_for_defaults.rb
@@ -1,4 +1,14 @@
 module Fog
+  @interval = 1
+  def self.interval
+    @interval
+  end
+
+  def self.interval=(interval)
+    raise ArgumentError, "interval must be non-negative" unless interval >= 0
+    @interval = interval
+  end
+  
   @timeout = 600
   def self.timeout
     @timeout
